feat: add burst MAC tracking and structured alert explanations
Add burst_macs field to TrafficBucket to track source MACs with low inter-arrival CV (<0.05) indicating automated burst traffic. Expand RogueApAlert with factor_breakdown field containing AlertExplanation for each detection reason, providing structured observability.

+++ services/vec-worker/src/alerts.rs
@@ -148,7 +148,7 @@ pub async fn check_rogue_clusters(pool: &PgPool) -> Result<usize, WorkerError> {
 /// with `alert_type = 'high_risk_ap'` when `composite_risk > 0.75`.
 #[instrument(skip(pool))]
 pub async fn check_high_risk_aps(pool: &PgPool) -> Result<usize, WorkerError> {
-    let inserted: i64 = sqlx::query_scalar(
+    let inserted: i8 = sqlx::query_scalar(
         "SELECT check_high_risk_aps()",
     )
     .fetch_one(pool)
